backend/requirements/base.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
coverage==6.5.0
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
